requirements.txt
pytest~=7.4.2
requests~=2.31.0
jsonschema~=4.19.0
selene==2.0.0rc4
allure-pytest
curlify~=2.2.1